Pre-release merge for version 0.2.4

+++ cookies/operations.py
@@ -5,6 +5,7 @@ from cookies.models import *
 from concepts.models import Concept
 
 import jsonpickle, datetime
+from itertools import groupby
 
 
 def add_creation_metadata(resource, user):
@@ -20,8 +21,56 @@ def add_creation_metadata(resource, user):
     })
 
 
+def _transfer_all_relations(from_instance, to_instance_id, content_type):
+    from_instance.relations_from.update(source_type=content_type,
+                                        source_instance_id=to_instance_id)
+    from_instance.relations_to.update(target_type=content_type,
+                                      target_instance_id=to_instance_id)
 
-def merge_conceptentities(entities, master_id=None):
+
+def prune_relations(resource):
+    """
+    Search for and remove duplicate relations for a :class:`.Resource`\.
+    """
+    value_type = ContentType.objects.get_for_model(Value)
+
+    def _search_and_destroy(relations):
+        def _delete_dupes(objs):    # objs is an iterator of values() dicts.
+            for obj in objs[1:]:    # Leave the first object.
+                Relation.objects.get(pk=obj[-1]).delete()
+
+        # We're looking for relations with the same predicate, whose
+        #  complementary object is of the same type and is either identical or
+        #  (if a Value) has the same value/content.
+        for pred, pr_relations in groupby(relations, lambda o: o[0]):
+            for ctype, ct_relations in groupby(pr_relations, lambda o: o[1]):
+                # We need to use this iterator twice, so we'll solidify
+                #  it as a list.
+                ct_relations = [o for o in ct_relations]
+
+                for iid, id_relations in groupby(ct_relations, lambda o: o[2]):
+                    _delete_dupes(list(id_relations)) # Target is precisely the same.
+
+                if ctype != value_type.id:    # Only applies to Value instances.
+                    continue
+
+                values = Value.objects.filter(pk__in=zip(*ct_relations)[2])\
+                            .order_by('id').values('id', '_value')
+
+                key = lambda *o: o[0][1]['_value']
+                for value, vl_relations in groupby(sorted(zip(ct_relations, values), key=key), key):
+                    v_relations = zip(*list(vl_relations))[0]
+                    _delete_dupes(v_relations)    # Target has the same value.
+
+    fields = ['predicate_id', 'target_type', 'target_instance_id', 'id']
+    _search_and_destroy(resource.relations_from.order_by(*fields).values_list(*fields))
+
+    fields = ['predicate_id', 'source_type', 'source_instance_id', 'id']
+    _search_and_destroy(resource.relations_to.order_by(*fields).values_list(*fields))
+
+
+
+def merge_conceptentities(entities, master_id=None, delete=True):
     """
     Merge :class:`.ConceptEntity` instances in the QuerySet ``entities``.
 
@@ -72,8 +121,40 @@ def merge_conceptentities(entities, master_id=None):
     # Update all Relations.
     to_merge = entities.filter(~Q(pk=master.id))
     for entity in to_merge:
-        entity.relations_from.update(source_type=conceptentity_type, source_instance_id=master.id)
-        entity.relations_to.update(target_type=conceptentity_type, target_instance_id=master.id)
+        _transfer_all_relations(entity, master.id, conceptentity_type)
 
-    to_merge.delete()
+    if delete:
+        to_merge.delete()
+    return master
+
+
+def merge_resources(resources, master_id=None, delete=True):
+    """
+    """
+    resource_type = ContentType.objects.get_for_model(Resource)
+    if resources.count() < 2:
+        raise RuntimeError("Need more than one Resource instance to merge")
+
+    with_content = resources.filter(content_resource=True)
+
+    if with_content.count() != 0 and with_content.count() != resources.count():
+        raise RuntimeError("Cannot merge content and non-content resources")
+
+    if master_id:
+        master = resources.get(pk=master_id)
+    else:
+        master = resources.first()
+
+    to_merge = resources.filter(~Q(pk=master.id))
+    for resource in to_merge:
+        _transfer_all_relations(resource, master.id, resource_type)
+        resource.content.all().update(for_resource=master)
+        for collection in resource.part_of.all():
+            master.part_of.add(collection)
+
+    prune_relations(master)
+
+    master.save()
+    if delete:
+        to_merge.delete()
     return master
